Create .gitkeep
This commit creates a .gitkeep file within the 'Slides/NFT-Basics' folder. This file ensures that the folder is tracked by Git even when empty. The folder will store slide decks related to NFT fundamentals, providing educational materials for veterinary professionals on the basics of NFTs, their utility, and the broader implications in blockchain technology. These slides will accompany future webinars and serve as a foundation for further learning on NFT technologies.