Create .gitkeep
This commit initializes the "Blockchain-Basics" folder under the "Resources" section of the repository. It adds a .gitkeep file to ensure the folder structure is maintained in the repository, allowing future educational materials related to blockchain technology to be added for veterinary professionals. This folder will serve as a storage space for external resources like articles, papers, and tools that support veterinarians in learning about blockchain fundamentals.