Create .gitkeep
This commit introduces the .gitkeep file to the Cardano-NFTs subfolder within the Guides section. The purpose of this file is to ensure the folder is tracked in Git, even when empty, aligning with the folder structure necessary for future guides on creating, managing, and understanding NFTs on the Cardano blockchain. This commit is part of the broader effort to create a comprehensive guide on blockchain and NFT education for veterinarians.