Create .gitkeep
Added .gitkeep file to ensure the NFT-Basics folder within the Guides section is tracked in the repository, allowing it to exist even if no content is present initially. This folder will serve as the location for introductory guides related to NFTs, including basic concepts and applications relevant to the veterinary blockchain education project. Further content will be progressively added as the project progresses, focusing on educating veterinarians on the foundational aspects of NFTs.